+++ node/src/types/kernel/api/storage/getRecentDocs/index.d.ts
@@ -17,6 +17,6 @@
 
 
 /* directories */
-export * as getRecentDocs from "./getRecentDocs";
 
 /* files */
+export * from "./response";