+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 更改日志 | Change Log
 
 - 更新 API `/api/file/readDir` | Update API `/api/file/readDir`
+- 添加 API `/api/storage/getRecentDocs` | Add API `/api/storage/getRecentDocs`
 
 ## v0.2.1 / 2023-08-11